safari-600.7-branch/LayoutTests/fast/shapes/parsing/parsing-shape-image-threshold-expected.txt

Test parsing of the CSS shape-image-threshold property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyleValue("-webkit-shape-image-threshold", "0") is "0"
PASS getComputedStyleValue("-webkit-shape-image-threshold", "0.5") is "0.5"
PASS getComputedStyleValue("-webkit-shape-image-threshold", "1") is "1"
PASS getComputedStyleValue("-webkit-shape-image-threshold", "-0.1") is "0"
PASS getComputedStyleValue("-webkit-shape-image-threshold", "1.1") is "1"
PASS getComputedStyleValue("-webkit-shape-image-threshold", "identifier") is "0"
PASS getComputedStyleValue("-webkit-shape-image-threshold", "'string'") is "0"
PASS getChildComputedStyle("-webkit-shape-image-threshold", "0", "0") is "0"
PASS getChildComputedStyle("-webkit-shape-image-threshold", "0", "1") is "1"
PASS getChildComputedStyle("-webkit-shape-image-threshold", "1", "-1") is "0"
PASS getChildComputedStyle("-webkit-shape-image-threshold", "2", "1") is "1"
PASS successfullyParsed is true

TEST COMPLETE
